fix to prevent generating term with free variable
If the replacing term were of different sort of the variable then the variable
was being produced, which meant that the instantiation rule would generate a
term with free variables, so the rule could be used in an unsound manner.

Now it produces none in such a case, preventing the issue.